backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
pydantic-settings>=2.1.0
anthropic>=0.18.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
httpx>=0.25.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
openpyxl>=3.1.2
pandas>=2.1.3
cryptography>=41.0.7
psycopg2-binary>=2.9.9
